+++ src/react/components/dropdown/XUIDropdown.js
@@ -229,6 +229,7 @@ export default class XUIDropdown extends PureComponent {
       fixedWidth,
       onCloseAnimationEnd,
       onOpenAnimationEnd,
+      onScroll,
       animateClosed,
       animateOpen,
       forceDesktop,
@@ -264,6 +265,7 @@ export default class XUIDropdown extends PureComponent {
           ignoreKeyboardEvents={ignoreKeyboardEvents}
           onHighlightChange={this.onHighlightChange}
           onKeyDown={this.keyDownHandler}
+          onScroll={onScroll}
           onSelect={onSelect}
           qaHook={qaHook}
           ref={this.panel}
@@ -314,6 +316,9 @@ XUIDropdown.propTypes = {
 
   /** Callback for adding additional onKeyPress functionality */
   onKeyDown: PropTypes.func,
+
+  /** Callback for adding additional onScroll functionality. */
+  onScroll: PropTypes.func,
 
   /** Whether focus should be restricted to the dropdown while it's open. */
   restrictFocus: PropTypes.bool,